+++ dcm4chee/dcm4chee-arc/trunk/dcm4jboss-web/src/java/org/dcm4chex/archive/web/maverick/ae/AEDelegate.java
@@ -128,9 +128,9 @@ public class AEDelegate {
 		public void updateAE( AEData ae, boolean checkHost) throws InstanceNotFoundException, MBeanException, ReflectionException {
 	        server.invoke(aeServiceName,
 	                "updateAE",
-	                new Object[]{ new Integer(ae.getPk()), ae.getTitle(), ae.getHostName(), 
+	                new Object[]{ new Long(ae.getPk()), ae.getTitle(), ae.getHostName(), 
 	        					new Integer(ae.getPort()),ae.getCipherSuitesAsString(), new Boolean(checkHost) },
-	                new String[]{ int.class.getName(), String.class.getName(), String.class.getName(),
+	                new String[]{ long.class.getName(), String.class.getName(), String.class.getName(),
 	        						int.class.getName(), String.class.getName(), boolean.class.getName()} );
 		}
