+++ dcm4chee/dcm4chee-arc/trunk/dcm4jboss-web/src/java/org/dcm4chex/archive/web/maverick/trash/ExpandTrashInstanceCtrl.java
@@ -52,9 +52,9 @@ import org.dcm4chex.archive.web.maverick.model.InstanceModel;
  */
 public class ExpandTrashInstanceCtrl extends ExpandTrashSeriesCtrl {
 
-    private int instancePk;
+    private long instancePk;
     
-    public final void setInstancePk(int pk) {
+    public final void setInstancePk(long pk) {
         this.instancePk = pk;
     }
